EVG-16546: Make commit chart collapsible
