+++ src/components/Accordion/index.tsx
@@ -82,13 +82,15 @@ const AccordionToggle = styled.span`
   }
 `;
 const AnimatedAccordion = styled.div`
-  max-height: 0;
   /* This is used to calculate a fixed height for the Accordion since height
-     transitions require a fixed height for their end height */
-  max-height: ${(props: { hide: boolean }): string => !props.hide && "6000px"};
-  overflow-y: ${(props: { hide: boolean }): string =>
-    props.hide ? "hidden" : "visible"};
-  transition: max-height 0.3s ease-in-out;
+      transitions require a fixed height for their end height */
+  max-height: ${(props: { hide: boolean }): string =>
+    props.hide ? "0px" : "9999px"};
+  overflow-y: ${(props: { hide: boolean }): string => props.hide && "hidden"};
+  transition: ${(props: { hide: boolean }): string =>
+    props.hide
+      ? "max-height 0.3s cubic-bezier(0, 1, 0, 1)"
+      : "max-height 0.6s ease-in-out"};
 `;
 const ContentsContainer = styled.div`
   margin-left: ${(props: { indent: boolean }): string =>
