+++ src/pages/commits/ActiveCommits/Grid.tsx
@@ -35,7 +35,7 @@ const DashedLine = styled.div`
   border: 1px dashed ${gray.light2};
 `;
 
-const SolidLine = styled.div`
+export const SolidLine = styled.div`
   width: 100%;
   border: 1px solid ${gray.light1};
 `;
